refactor: organize components into separate folders
