+++ src/app/icons/ClearSkyNightIcon.tsx
