+++ src/app/icons/PartlyCloudyDayIcon.tsx
