+++ src/app/icons/PartlyCloudyNightIcon.tsx
