+++ src/app/icons/SnowIcon.tsx
