+++ src/app/icons/ThunderStormIcon.tsx
