+++ src/app/ui/AnalogClock/AnalogClock.stories.ts
