+++ src/app/ui/DigitalClock/index.ts
