+++ src/app/ui/WeatherForecast/TemperatureBar.tsx
